+++ db/migrate/029_create_sums.rb
@@ -1,13 +1,16 @@
 class CreateSums < ActiveRecord::Migration
   def self.up
-    create_table :sums do |t|
-	t.column :key,		:string
-	t.column :amount,	:float,	:default	=>	0
-	t.timestamps
-    end
+	begin
+		create_table :sums do |t|
+			t.column :key,		:string
+			t.column :amount,	:float,	:default	=>	0
+			t.timestamps
+		end
+	rescue
+	end
   end
 
   def self.down
-    drop_table :sums
+    drop_table :sums	if	ENV['RAILS_ENV'] == 'production'
   end
 end
